Updates for new balsam API changes
