+++ parsl/app/balsam.py
@@ -6,6 +6,110 @@ from parsl.app.errors import wrap_error
 from parsl.app.app import AppBase
 from parsl.dataflow.dflow import DataFlowKernelLoader
 
+from balsam.site import ApplicationDefinition
+
+
+class BashRunner(ApplicationDefinition):
+    """
+    Balsam Bash Runner. P
+    """
+    environment_variables = {}
+    command_template = '/bin/bash -c {{ command }}'
+    parameters = {}
+    transfers = {}
+    
+    site = 0
+
+    def postprocess(self):
+        from balsam.api import site_config
+
+        workdir = self.job.resolve_workdir(site_config.data_path)
+        stdout = workdir.joinpath("job.out").read_text().strip()
+        self.job.data = {'result': stdout, 'type': 'bash'}
+        self.job.state = "POSTPROCESSED"
+
+    def handle_timeout(self):
+        self.job.state = "RESTART_READY"
+
+    def handle_error(self):
+        self.job.state = "FAILED"
+
+class AppRunner(ApplicationDefinition):
+    """
+    Balsam App Runner for python apps
+    """
+    environment_variables = {}
+    command_template = '{{ python }} app.py'
+    parameters = {}
+    transfers = {}
+
+    site = 0
+
+    def postprocess(self):
+        from balsam.api import site_config
+        import json
+
+        workdir = self.job.resolve_workdir(site_config.data_path)
+        print('WORKDIR: ', workdir)
+        stdout = workdir.joinpath("job.out").read_text().strip()
+        metadata = workdir.joinpath("job.metadata").read_text().strip()
+        print('STDOUT: ', stdout)
+
+        print("METADATA: ", metadata)
+        try:
+            metadata = json.loads(metadata)
+            self.job.data = metadata
+        except:
+            import traceback
+            print(traceback.format_exc())
+            self.job.data = {'result': stdout, 'type': 'bash'}
+
+        self.job.state = "POSTPROCESSED"
+
+    def handle_timeout(self):
+        self.job.state = "RESTART_READY"
+
+    def handle_error(self):
+        self.job.state = "FAILED"
+        
+class ContainerRunner(ApplicationDefinition):
+    """
+    Container Runner. 
+    """
+    environment_variables = {}
+    command_template = 'singularity exec --bind {{ datadir }}:/data --bind .:/app {{ image }} python /app/app.py'
+    parameters = {}
+    transfers = {}
+
+    site = 0
+
+    def postprocess(self):
+        from balsam.api import site_config
+        import json
+
+        workdir = self.job.resolve_workdir(site_config.data_path)
+        print('WORKDIR: ', workdir)
+        stdout = workdir.joinpath("job.out").read_text().strip()
+        metadata = workdir.joinpath("job.metadata").read_text().strip()
+        print('STDOUT: ', stdout)
+
+        print("METADATA: ", metadata)
+        try:
+            metadata = json.loads(metadata)
+            self.job.data = metadata
+        except:
+            import traceback
+            print(traceback.format_exc())
+            self.job.data = {'result': stdout, 'type': 'bash'}
+
+        self.job.state = "POSTPROCESSED"
+
+    def handle_timeout(self):
+        self.job.state = "RESTART_READY"
+
+    def handle_error(self):
+        self.job.state = "FAILED"
+
 
 class BalsamApp(AppBase):
